break out session_metrics
