+++ client/session.go
@@ -41,11 +41,12 @@ import (
 	"github.com/m3db/m3db/topology"
 	"github.com/m3db/m3db/ts"
 	xio "github.com/m3db/m3db/x/io"
+
 	xerrors "github.com/m3db/m3x/errors"
-	"github.com/m3db/m3x/log"
+	xlog "github.com/m3db/m3x/log"
 	"github.com/m3db/m3x/pool"
-	"github.com/m3db/m3x/retry"
-	"github.com/m3db/m3x/sync"
+	xretry "github.com/m3db/m3x/retry"
+	xsync "github.com/m3db/m3x/sync"
 	xtime "github.com/m3db/m3x/time"
 
 	"github.com/uber-go/tally"
@@ -113,37 +114,6 @@ type session struct {
 	streamBlocksMetadataBatchTimeout time.Duration
 	streamBlocksBatchTimeout         time.Duration
 	metrics                          sessionMetrics
-}
-
-type sessionMetrics struct {
-	sync.RWMutex
-	writeSuccess               tally.Counter
-	writeErrors                tally.Counter
-	writeNodesRespondingErrors []tally.Counter
-	fetchSuccess               tally.Counter
-	fetchErrors                tally.Counter
-	fetchNodesRespondingErrors []tally.Counter
-	streamFromPeersMetrics     map[uint32]streamFromPeersMetrics
-}
-
-func newSessionMetrics(scope tally.Scope) sessionMetrics {
-	return sessionMetrics{
-		writeSuccess:           scope.Counter("write.success"),
-		writeErrors:            scope.Counter("write.errors"),
-		fetchSuccess:           scope.Counter("fetch.success"),
-		fetchErrors:            scope.Counter("fetch.errors"),
-		streamFromPeersMetrics: make(map[uint32]streamFromPeersMetrics),
-	}
-}
-
-type streamFromPeersMetrics struct {
-	fetchBlocksFromPeers      tally.Gauge
-	metadataFetches           tally.Gauge
-	metadataFetchBatchCall    tally.Counter
-	metadataFetchBatchSuccess tally.Counter
-	metadataFetchBatchError   tally.Counter
-	metadataReceived          tally.Counter
-	blocksEnqueueChannel      tally.Gauge
 }
 
 type newHostQueueFn func(
@@ -250,40 +220,6 @@ func (s *session) streamFromPeersMetricsForShard(shard uint32) *streamFromPeersM
 	s.metrics.streamFromPeersMetrics[shard] = m
 	s.metrics.Unlock()
 	return &m
-}
-
-func (s *session) incWriteMetrics(consistencyResultErr error, respErrs int32) {
-	if idx := s.nodesRespondingErrorsMetricIndex(respErrs); idx >= 0 {
-		s.metrics.writeNodesRespondingErrors[idx].Inc(1)
-	}
-	if consistencyResultErr == nil {
-		s.metrics.writeSuccess.Inc(1)
-	} else {
-		s.metrics.writeErrors.Inc(1)
-	}
-}
-
-func (s *session) incFetchMetrics(consistencyResultErr error, respErrs int32) {
-	if idx := s.nodesRespondingErrorsMetricIndex(respErrs); idx >= 0 {
-		s.metrics.fetchNodesRespondingErrors[idx].Inc(1)
-	}
-	if consistencyResultErr == nil {
-		s.metrics.fetchSuccess.Inc(1)
-	} else {
-		s.metrics.fetchErrors.Inc(1)
-	}
-}
-
-func (s *session) nodesRespondingErrorsMetricIndex(respErrs int32) int32 {
-	idx := respErrs - 1
-	replicas := int32(s.Replicas())
-	if respErrs > replicas {
-		// Cap to the max replicas, we might get more errors
-		// when a node is initializing a shard causing replicas + 1
-		// nodes to respond to operations
-		idx = replicas - 1
-	}
-	return idx
 }
 
 func (s *session) Open() error {
